stockcharts/requirements.txt
Django
djangorestframework
psycopg2-binary
kafka-python
python-dotenv
requests
pandas
alpaca-trade-api 
plotly